requirements.txt
python-dotenv
pymupdf
fuzzywuzzy
openai
streamlit
# elevenlabs
# langchain
nltk
scikit-learn
sentence-transformers

